# CommandService/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CommandService",
    "version": "v1"
  },
  "paths": {
    "/api/c/platforms/{platformId}/Commands": {
      "get": {
        "tags": [
          "Commands"
        ],
        "parameters": [
          {
            "name": "platformId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandReadDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Commands"
        ],
        "parameters": [
          {
            "name": "platformId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommandCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/c/platforms/{platformId}/Commands/{commandId}": {
      "get": {
        "tags": [
          "Commands"
        ],
        "operationId": "GetCommandForPlatform",
        "parameters": [
          {
            "name": "platformId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "commandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/c/Platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformReadDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Platforms"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CommandReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "howTo": {
            "type": "string",
            "nullable": true
          },
          "commandLine": {
            "type": "string",
            "nullable": true
          },
          "platformId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CommandCreateDto": {
        "required": [
          "commandLine",
          "howTo"
        ],
        "type": "object",
        "properties": {
          "howTo": {
            "type": "string"
          },
          "commandLine": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PlatformReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}